requirements.txt
# Dependências principais
google-generativeai>=0.8.3
python-dotenv>=1.0.0
aiohttp>=3.9.0

# Dependências para integrações
slack-sdk>=3.0.0
PyGithub>=2.0.0
jira>=3.0.0
discord.py>=2.0.0

# Dependências para desenvolvimento
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.0.0
flake8>=6.0.0
mypy>=1.0.0
sphinx>=7.0.0
sphinx-rtd-theme>=1.0.0
